requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
easydict==1.10
fire==0.4.0
frozenlist==1.3.3
fsspec==2022.10.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
networkx==2.8.8
numpy==1.20.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytorch-lightning==1.8.0.post1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
Shapely==1.8.5.post1
six==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
tifffile==2022.10.10
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.10.2
torchvision==0.14.0
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
